% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{colors_ifau}
\alias{colors_ifau}
\title{IFAU colors}
\usage{
colors_ifau(n = 7)
}
\arguments{
\item{n}{Number of colors}
}
\description{
Return a vector of the first \code{n} colors from the IFAU color scheme.
}
