% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifau_save.R
\name{ifau_save}
\alias{ifau_save}
\title{Export graph to file}
\usage{
ifau_save(
  filename,
  plot = last_plot(),
  device = NULL,
  path = NULL,
  scale = 1,
  width = 12.4,
  height = 6.9,
  units = "cm",
  dpi = 300,
  limitsize = TRUE,
  bg = NULL,
  create.dir = FALSE,
  ...
)
}
\description{
Export graph to file
}
