% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{linetypes_ifau}
\alias{linetypes_ifau}
\title{IFAU linetypes}
\usage{
linetypes_ifau(n = 7)
}
\arguments{
\item{n}{Number of linetypes}
}
\description{
Return a vector of the first \code{n} line patterns from the IFAU graph scheme.
}
