% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{linetypes_ifau_fun}
\alias{linetypes_ifau_fun}
\title{IFAU linetypes function}
\usage{
linetypes_ifau_fun()
}
\description{
Return the \code{linetypes_ifau} function, for use when defining custom linetype scales.
}
