% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_linetype_ifau}
\alias{scale_linetype_ifau}
\title{IFAU linetype scale}
\usage{
scale_linetype_ifau(...)
}
\arguments{
\item{...}{Arguments passed on to \code{discrete_scale}}
}
\description{
Use IFAU patterns for \code{ggplot2} lines.
}
