% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{shapes_ifau}
\alias{shapes_ifau}
\title{IFAU shapes}
\usage{
shapes_ifau(n = 7)
}
\arguments{
\item{n}{Number of shapes}
}
\description{
Return a vector of the first \code{n} point market shapes from the IFAU graph scheme.
}
