% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifau_theme.R
\name{theme_ifau_minimal}
\alias{theme_ifau_minimal}
\title{Minimal IFAU theme}
\usage{
theme_ifau_minimal()
}
\description{
Sets plot aspects text, axes, grid, and legend. Does not set colors, linetypes, shapes etc.
}
